@@ -77,41 +73,40 @@
     "@babel/core": "^7.11.6",
     "@babel/eslint-parser": "^7.16.5",
     "@babel/preset-env": "^7.11.5",
-    "@playwright/test": "^1.18.1",
+    "@eslint/js": "^8.43.0",
+    "@playwright/test": "^1.39.0",
+    "@rollup/plugin-commonjs": "^25.0.2",
+    "@rollup/plugin-node-resolve": "^15.1.0",
+    "@rollup/plugin-replace": "^5.0.2",
+    "@rollup/plugin-terser": "^0.4.3",
+    "@types/eslint": "^8.40.2",
     "autoprefixer-stylus": "^1.0.0",
-    "axios": "^0.21.1",
+    "axios": "^1.5.0",
     "babel-jest": "^27.4.6",
     "browser-sync": "^2.26.12",
     "chokidar": "^3.4.2",
     "common-tags": "^1.8.0",
-    "conventional-changelog-cli": "^2.1.0",
+    "conventional-changelog-cli": "^3.0.0",
     "copy-dir": "^1.2.0",
-    "cross-env": "^6.0.3",
+    "cross-env": "^7.0.3",
     "cssnano": "^4.1.10",
-    "eslint": "^8.7.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^26.0.0",
-    "eslint-plugin-playwright": "^0.8.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "husky": "^3.1.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.2",
+    "eslint-plugin-playwright": "^0.16.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "globals": "^13.20.0",
+    "husky": "^8.0.3",
     "jest": "^27.4.7",
-    "lerna": "^3.22.1",
-    "lint-staged": "^10.4.0",
-    "live-server": "^1.2.1",
-    "mkdirp": "^0.5.1",
+    "lint-staged": "^13.2.2",
+    "mkdirp": "^3.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
+    "prettier": "^2.8.8",
     "rimraf": "^3.0.0",
-    "rollup": "^1.23.1",
-    "rollup-plugin-async": "^1.2.0",
-    "rollup-plugin-buble": "^0.19.8",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-uglify": "^6.0.4",
+    "rollup": "^3.25.3",
     "serve-handler": "^6.1.2",
-    "stylus": "^0.54.5",
+    "stylus": "^0.60.0",
     "vue2": "npm:vue@^2.6.12",
     "vue3": "npm:vue@^3.0.0",
     "xhr-mock": "^2.5.1"
