@@ -1,81 +1,7 @@
-// Modules, constants, and variables
-// -----------------------------------------------------------------------------
-const docsifyInit = require('../helpers/docsify-init');
-const { test, expect } = require('./fixtures/docsify-init-fixture');
+import docsifyInit from '../helpers/docsify-init.js';
+import { test, expect } from './fixtures/docsify-init-fixture.js';
 
-// Suite
-// -----------------------------------------------------------------------------
-test.describe('Example Tests', () => {
-  // Tests
-  // ---------------------------------------------------------------------------
-  test('dom manipulation', async ({ page }) => {
-    const testText = 'This is a test';
-    const testHTML = `<h1>Test</h1><p>${testText}</p>`;
-
-    // Inject HTML
-    await page.setContent(testHTML);
-
-    // Get reference to page element
-    const bodyElm = page.locator('body');
-    const pElm = page.locator('body > p');
-
-    // Add class to element and test
-    await bodyElm.evaluate(elm => elm.classList.add('foo'));
-
-    // Tests
-    await expect(bodyElm).toHaveClass('foo');
-    await expect(bodyElm).toContainText('Test');
-    await expect(pElm).toHaveCount(1);
-    await expect(pElm).toHaveText(testText);
-    await expect(pElm).not.toHaveText('NOPE');
-  });
-
-  test('javascript in browser context', async ({ page }) => {
-    // Get native DOM values
-    const clientDimensions = await page.evaluate(() => {
-      return {
-        width: document.documentElement.clientWidth,
-        height: document.documentElement.clientHeight,
-      };
-    });
-
-    expect(clientDimensions).toHaveProperty('width');
-    expect(typeof clientDimensions.width).toBe('number');
-    expect(clientDimensions).toHaveProperty('height');
-    expect(typeof clientDimensions.height).toBe('number');
-
-    // Get result of script executed in browser context
-    const scriptResult = await page.evaluate(
-      numbers => {
-        const result = numbers.reduce(
-          (accumulator, currentValue) => accumulator + currentValue
-        );
-
-        return Promise.resolve(result);
-      },
-      [1, 2, 3]
-    );
-
-    expect(scriptResult).toBe(6);
-
-    // Get result of local function executed in browser context
-    function add(...addends) {
-      return addends.reduce(
-        (accumulator, currentValue) => accumulator + currentValue
-      );
-    }
-
-    const functionResult = await page.evaluate(`
-      const add = ${add.toString()};
-
-      const result = add(1, 2, 3);
-
-      Promise.resolve(result);
-    `);
-
-    expect(functionResult).toBe(6);
-  });
-
+test.describe('Creating a Docsify site (e2e tests in Playwright)', () => {
   test('manual docsify site using playwright methods', async ({ page }) => {
     // Add docsify target element
     await page.setContent('<div id="app"></div>');
@@ -84,7 +10,6 @@ test.describe('Example Tests', () => {
     await page.evaluate(() => {
       window.$docsify = {
         el: '#app',
-        basePath: '/docs/',
         themeColor: 'red',
       };
     });
@@ -110,9 +35,6 @@ test.describe('Example Tests', () => {
     // Load custom docsify
     // (See ./helpers/docsifyInit.js for details)
     await docsifyInit({
-      config: {
-        basePath: '/docs/',
-      },
       // _logHTML: true,
     });
 
