@@ -1,17 +1,9 @@
-const { waitForFunction, waitForText } = require('../helpers/wait-for');
+import { waitForFunction, waitForText } from '../helpers/wait-for.js';
+import docsifyInit from '../helpers/docsify-init.js';
 
-const docsifyInit = require('../helpers/docsify-init');
-
-// Suite
-// -----------------------------------------------------------------------------
-describe('Example Tests', function () {
-  // Tests
-  // ---------------------------------------------------------------------------
+describe('Creating a Docsify site (integration tests in Jest)', function () {
   test('Docsify /docs/ site using docsifyInit()', async () => {
     await docsifyInit({
-      config: {
-        basePath: '/docs/',
-      },
       // _logHTML: true,
     });
 
