@@ -1,18 +1,17 @@
-const stripIndent = require('common-tags/lib/stripIndent');
-const docsifyInit = require('../helpers/docsify-init');
+import stripIndent from 'common-tags/lib/stripIndent/index.js';
+import docsifyInit from '../helpers/docsify-init.js';
+import { waitForText } from '../helpers/wait-for.js';
 
 // Suite
 // -----------------------------------------------------------------------------
 describe('render', function () {
-  // Setup & Teardown
-  // -------------------------------------------------------------------------
-  beforeEach(async () => {
-    await docsifyInit();
-  });
-
   // Helpers
   // ---------------------------------------------------------------------------
   describe('helpers', () => {
+    beforeEach(async () => {
+      await docsifyInit();
+    });
+
     test('important content', () => {
       const output = window.marked('!> Important content');
 
@@ -33,6 +32,10 @@ describe('render', function () {
   // Lists
   // ---------------------------------------------------------------------------
   describe('lists', function () {
+    beforeEach(async () => {
+      await docsifyInit();
+    });
+
     test('as unordered task list', async function () {
       const output = window.marked(stripIndent`
         - [x] Task 1
@@ -100,11 +103,15 @@ describe('render', function () {
   // Images
   // ---------------------------------------------------------------------------
   describe('images', function () {
+    beforeEach(async () => {
+      await docsifyInit();
+    });
+
     test('regular', async function () {
       const output = window.marked('![alt text](http://imageUrl)');
 
       expect(output).toMatchInlineSnapshot(
-        `"<p><img src=\\"http://imageUrl\\" data-origin=\\"http://imageUrl\\" alt=\\"alt text\\"></p>"`
+        `"<p><img src=\\"http://imageUrl\\" data-origin=\\"http://imageUrl\\" alt=\\"alt text\\"  /></p>"`
       );
     });
 
@@ -136,35 +143,37 @@ describe('render', function () {
       );
     });
 
-    describe('size', function () {
-      test('width and height', async function () {
-        const output = window.marked(
-          "![alt text](http://imageUrl ':size=WIDTHxHEIGHT')"
-        );
+    test('width and height', async function () {
+      const output = window.marked(
+        "![alt text](http://imageUrl ':size=WIDTHxHEIGHT')"
+      );
 
-        expect(output).toMatchInlineSnapshot(
-          `"<p><img src=\\"http://imageUrl\\" data-origin=\\"http://imageUrl\\" alt=\\"alt text\\" width=\\"WIDTH\\" height=\\"HEIGHT\\" /></p>"`
-        );
-      });
+      expect(output).toMatchInlineSnapshot(
+        `"<p><img src=\\"http://imageUrl\\" data-origin=\\"http://imageUrl\\" alt=\\"alt text\\" width=\\"WIDTH\\" height=\\"HEIGHT\\" /></p>"`
+      );
+    });
 
-      test('width', async function () {
-        const output = window.marked("![alt text](http://imageUrl ':size=50')");
+    test('width', async function () {
+      const output = window.marked("![alt text](http://imageUrl ':size=50')");
 
-        expect(output).toMatchInlineSnapshot(
-          `"<p><img src=\\"http://imageUrl\\" data-origin=\\"http://imageUrl\\" alt=\\"alt text\\" width=\\"50\\" /></p>"`
-        );
-      });
+      expect(output).toMatchInlineSnapshot(
+        `"<p><img src=\\"http://imageUrl\\" data-origin=\\"http://imageUrl\\" alt=\\"alt text\\" width=\\"50\\" /></p>"`
+      );
     });
   });
 
   // Headings
   // ---------------------------------------------------------------------------
   describe('headings', function () {
+    beforeEach(async () => {
+      await docsifyInit();
+    });
+
     test('h1', async function () {
       const output = window.marked('# h1 tag');
 
       expect(output).toMatchInlineSnapshot(
-        `"<h1 id=\\"h1-tag\\"><a href=\\"#/?id=h1-tag\\" data-id=\\"h1-tag\\" class=\\"anchor\\"><span>h1 tag</span></a></h1>"`
+        `"<h1 id=\\"h1-tag\\" tabindex=\\"-1\\"><a href=\\"#/?id=h1-tag\\" data-id=\\"h1-tag\\" class=\\"anchor\\"><span>h1 tag</span></a></h1>"`
       );
     });
 
@@ -172,7 +181,7 @@ describe('render', function () {
       const output = window.marked('## h2 tag');
 
       expect(output).toMatchInlineSnapshot(
-        `"<h2 id=\\"h2-tag\\"><a href=\\"#/?id=h2-tag\\" data-id=\\"h2-tag\\" class=\\"anchor\\"><span>h2 tag</span></a></h2>"`
+        `"<h2 id=\\"h2-tag\\" tabindex=\\"-1\\"><a href=\\"#/?id=h2-tag\\" data-id=\\"h2-tag\\" class=\\"anchor\\"><span>h2 tag</span></a></h2>"`
       );
     });
 
@@ -180,7 +189,7 @@ describe('render', function () {
       const output = window.marked('### h3 tag');
 
       expect(output).toMatchInlineSnapshot(
-        `"<h3 id=\\"h3-tag\\"><a href=\\"#/?id=h3-tag\\" data-id=\\"h3-tag\\" class=\\"anchor\\"><span>h3 tag</span></a></h3>"`
+        `"<h3 id=\\"h3-tag\\" tabindex=\\"-1\\"><a href=\\"#/?id=h3-tag\\" data-id=\\"h3-tag\\" class=\\"anchor\\"><span>h3 tag</span></a></h3>"`
       );
     });
 
@@ -188,7 +197,7 @@ describe('render', function () {
       const output = window.marked('#### h4 tag');
 
       expect(output).toMatchInlineSnapshot(
-        `"<h4 id=\\"h4-tag\\"><a href=\\"#/?id=h4-tag\\" data-id=\\"h4-tag\\" class=\\"anchor\\"><span>h4 tag</span></a></h4>"`
+        `"<h4 id=\\"h4-tag\\" tabindex=\\"-1\\"><a href=\\"#/?id=h4-tag\\" data-id=\\"h4-tag\\" class=\\"anchor\\"><span>h4 tag</span></a></h4>"`
       );
     });
 
@@ -196,7 +205,7 @@ describe('render', function () {
       const output = window.marked('##### h5 tag');
 
       expect(output).toMatchInlineSnapshot(
-        `"<h5 id=\\"h5-tag\\"><a href=\\"#/?id=h5-tag\\" data-id=\\"h5-tag\\" class=\\"anchor\\"><span>h5 tag</span></a></h5>"`
+        `"<h5 id=\\"h5-tag\\" tabindex=\\"-1\\"><a href=\\"#/?id=h5-tag\\" data-id=\\"h5-tag\\" class=\\"anchor\\"><span>h5 tag</span></a></h5>"`
       );
     });
 
@@ -204,12 +213,18 @@ describe('render', function () {
       const output = window.marked('###### h6 tag');
 
       expect(output).toMatchInlineSnapshot(
-        `"<h6 id=\\"h6-tag\\"><a href=\\"#/?id=h6-tag\\" data-id=\\"h6-tag\\" class=\\"anchor\\"><span>h6 tag</span></a></h6>"`
+        `"<h6 id=\\"h6-tag\\" tabindex=\\"-1\\"><a href=\\"#/?id=h6-tag\\" data-id=\\"h6-tag\\" class=\\"anchor\\"><span>h6 tag</span></a></h6>"`
       );
     });
   });
 
+  // Links
+  // ---------------------------------------------------------------------------
   describe('link', function () {
+    beforeEach(async () => {
+      await docsifyInit();
+    });
+
     test('regular', async function () {
       const output = window.marked('[alt text](http://url)');
 
@@ -270,6 +285,74 @@ describe('render', function () {
       expect(output).toMatchInlineSnapshot(
         `"<p><a href=\\"http://url\\" target=\\"_blank\\"  rel=\\"noopener\\" id=\\"someCssID\\">alt text</a></p>"`
       );
+    });
+  });
+
+  // Skip Link
+  // ---------------------------------------------------------------------------
+  describe('skip link', () => {
+    test('renders default skip link and label', async () => {
+      await docsifyInit();
+
+      const elm = document.getElementById('skip-to-content');
+      const expectText = 'Skip to main content';
+
+      expect(elm.textContent).toBe(expectText);
+      expect(elm.outerHTML).toMatchInlineSnapshot(
+        `"<button id=\\"skip-to-content\\">Skip to main content</button>"`
+      );
+    });
+
+    test('renders custom label from config string', async () => {
+      const expectText = 'test';
+
+      await docsifyInit({
+        config: {
+          skipLink: expectText,
+        },
+      });
+
+      const elm = document.getElementById('skip-to-content');
+
+      expect(elm.textContent).toBe(expectText);
+    });
+
+    test('renders custom label from config object', async () => {
+      const getSkipLinkText = () =>
+        document.getElementById('skip-to-content').textContent;
+
+      await docsifyInit({
+        config: {
+          skipLink: {
+            '/dir1/dir2/': 'baz',
+            '/dir1/': 'bar',
+          },
+        },
+      });
+
+      window.location.hash = '/dir1/dir2/';
+      await waitForText('#skip-to-content', 'baz');
+      expect(getSkipLinkText()).toBe('baz');
+
+      window.location.hash = '/dir1/';
+      await waitForText('#skip-to-content', 'bar');
+      expect(getSkipLinkText()).toBe('bar');
+
+      // Fallback to default
+      window.location.hash = '';
+      await waitForText('#skip-to-content', 'Skip to main content');
+      expect(getSkipLinkText()).toBe('Skip to main content');
+    });
+
+    test('does not render skip link when false', async () => {
+      await docsifyInit({
+        config: {
+          skipLink: false,
+        },
+      });
+      const elm = document.getElementById('skip-to-content') || false;
+
+      expect(elm).toBe(false);
     });
   });
 });
