@@ -1,4 +1,4 @@
-import * as dom from '../util/dom';
+import * as dom from '../util/dom.js';
 
 let barEl;
 let timeId;
@@ -10,19 +10,25 @@ function init() {
   const div = dom.create('div');
 
   div.classList.add('progress');
+  div.setAttribute('role', 'progressbar');
+  div.setAttribute('aria-valuemin', '0');
+  div.setAttribute('aria-valuemax', '100');
+  div.setAttribute('aria-label', 'Loading...');
   dom.appendTo(dom.body, div);
   barEl = div;
 }
 
 /**
  * Render progress bar
+ * @param {{step: number, loaded?: undefined, total?: undefined} | {step?: undefined, loaded: number, total: number}} info
  */
-export default function ({ loaded, total, step }) {
+export default function (info) {
+  const { loaded, total, step } = info;
   let num;
 
   !barEl && init();
 
-  if (step) {
+  if (typeof step !== 'undefined') {
     num = parseInt(barEl.style.width || 0, 10) + step;
     num = num > 80 ? 80 : num;
   } else {
@@ -31,6 +37,7 @@ export default function ({ loaded, total, step }) {
 
   barEl.style.opacity = 1;
   barEl.style.width = num >= 95 ? '100%' : num + '%';
+  barEl.setAttribute('aria-valuenow', num >= 95 ? 100 : num);
 
   if (num >= 95) {
     clearTimeout(timeId);
@@ -38,6 +45,7 @@ export default function ({ loaded, total, step }) {
     timeId = setTimeout(_ => {
       barEl.style.opacity = 0;
       barEl.style.width = '0%';
+      barEl.removeAttribute('aria-valuenow');
     }, 200);
   }
 }
