@@ -1,9 +1,11 @@
 /**
- * Create a cached version of a pure function.
+ * Create a cached version of fn that given an input string returns a
+ * cached return value mapped from the string, regardless if fn is a new function every time.
+ * created.
  * @param {*} fn The function call to be cached
  * @void
  */
-
+// TODO Replace this with a proper memo(fn) based on args per function.
 export function cached(fn) {
   const cache = Object.create(null);
   return function (str) {
@@ -20,33 +22,10 @@ export const hyphenate = cached(str => {
   return str.replace(/([A-Z])/g, m => '-' + m.toLowerCase());
 });
 
-export const hasOwn = Object.prototype.hasOwnProperty;
-
-/**
- * Simple Object.assign polyfill
- * @param {Object} to The object to be merged with
- * @returns {Object} The merged object
- */
-export const merge =
-  Object.assign ||
-  function (to) {
-    for (let i = 1; i < arguments.length; i++) {
-      const from = Object(arguments[i]);
-
-      for (const key in from) {
-        if (hasOwn.call(from, key)) {
-          to[key] = from[key];
-        }
-      }
-    }
-
-    return to;
-  };
-
 /**
  * Check if value is primitive
  * @param {*} value Checks if a value is primitive
- * @returns {Boolean} Result of the check
+ * @returns {value is string | number} Result of the check
  */
 export function isPrimitive(value) {
   return typeof value === 'string' || typeof value === 'number';
@@ -61,7 +40,7 @@ export function noop() {}
 /**
  * Check if value is function
  * @param {*} obj Any javascript object
- * @returns {Boolean} True if the passed-in value is a function
+ * @returns {obj is Function} True if the passed-in value is a function
  */
 export function isFn(obj) {
   return typeof obj === 'function';
@@ -94,6 +73,9 @@ export function isExternal(url) {
       ''
     ) !== location.host
   ) {
+    return true;
+  }
+  if (/^\/\\/.test(url)) {
     return true;
   }
   return false;
